// src/lib/types.ts
export type Student = {
  nom: string;
  postnom: string;
  prenom: string;
  inspiration: string;
  impression: string;
  reve: string;
  email: string;
  numero: string;
  bases: boolean;
  bureautique?: boolean | undefined;
  mobile?: boolean | undefined;
  web?: boolean | undefined;
  desktop?: boolean | undefined;
  video?: boolean | undefined;
  photo?: boolean | undefined;
};